Add test for complex tracing access chains.
DXC seems to always separate allocations so this "happens to work"
without us having to do anything else.

No idea if this is part of the spec that alloca'd variables must be
directly used in a trace and that the trace has a unique usage type
w.r.t. payload kind.