Handle "odd" integer sizes.
LLVM IR supports "weird" integer sizes, and it implies that backend
deals with masking where appropriate.

In contexts where sign matters, we need to extend the input to physical
bitwidth from logical bitwidth.

For now, this is probably not exhaustive, but handle trunc/zext/sext and
some arithmetic opcodes which are sign sensitive.