Add test for Gather with 4 offsets
DXC ends up producing 4 texture gathers here, and these are called out as being potentially variable in the HLSL spec.

In the const case, we could fold this to ConstOffsets in the future but it might be better to let the driver do that as it's a pretty easy pattern.