Avoid merging to continue block in certain "impossible" scenarios.
Strictly speaking, we don't have to merge to continue target like our
lives depended on it. In most cases, we can just branch to continue
target without any specific merge, but we'd like to be maximally
convergent as much as we can. For now, maintain a heuristic for this
where we try really hard to merge and back off if we have to.