Avoid potential unstable codegen with add_ids.
These arguments can be processed out of order leading to unstable
codegen. We haven't run into a concrete issue with this yet, but be
defensive.

add_ids is okay if there is at most one call in the list. Should
probably remove the add_ids altogether, but a lot of churn.