Handle transposed loops.
It is possible that a loop breaks out a few levels while the continuing
path after the loop breaks out of more levels.

This is impossible to represent normally.

We have to transpose the loop such that the breaking construct is the
merge point, and the continuing path becomes the breaking construct
instead.

This obliterates the CFG after retargeting branches like this, so we
have to recompute the CFG before doing anything else.

Ridiculously complicated ... :(