Serialize interleaved merge scopes.
In heavily unrolled loops with many different breaks, we can end up with
code that is impossible to express with normal structured control flow
without deinterleaving the breaks.

The strategy is to pull all break blocks to a merge block which
switch-dispatches to the appropriate break construct, then reconvenes at
the merge block.