Use a more proper demangler.
It does not seem to be a requirement that valid C++ identifiers are
used in the demangled name. Find a substring between ? and @, which
seems to be based on MSVC mangling rules and has been stable since forever,
so this should be fine.

Add an API to query the demangled name as well for convenience.