Avoid false loops in merge block splits.
It's possible to have a diamond merge pattern where

A -> B (B = A.succ helper block)
A is fake loop header merging to C
B is header merging to D
C and D are not reachable
C and D merge to E.

During loop merge, there is risk of the outer loop scope of A -> C
triggering branches into the B -> D scope when rewriting ladder blocks.

Block branch rewrites if the block is a ladder block for the outer
scope.