Optimize certain bad CBV -> alloca copy patterns.
Tremendous uplifts in Persona 3 Reload in certain scenes.

Implement this by tracking all writes to an alloca() buffer.
If we're writing with pattern of alloca[index] = CBV + constant + stride * index,
we can translate GEP during instruction emit to directly access CBV.

We currently don't dead-load eliminate the CBV loads. Possible in theory
to do it, but it's not that easy since dxil-spirv isn't a full scale
compiler. We expect drivers to do this for us.